requirements.txt
asgiref==3.5.2
backports.weakref==1.0.post1
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==3.2.16
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==7.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt-4==1.12.0
djangorestframework-simplejwt==5.2.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.4
phonenumbers==8.12.56
Pillow==9.2.0
psycopg2-binary==2.9.5
PyJWT==2.0.0
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.4
qrcode==7.3.1
requests==2.28.1
sqlparse==0.4.3
twilio==7.14.2
urllib3==1.26.12
waitress==2.1.2
whitenoise==6.2.0
